feat(#55): Division feed en plusieurs classes et utilisation d'une factory
