+++ divertech/app/src/main/java/Si3/divertech/MainActivity.java
@@ -26,11 +26,12 @@ public class MainActivity extends AppCompatActivity implements ClickableActivity
             button.setVisibility(View.VISIBLE);
             button.setOnClickListener(click -> finish());
         }
+
+        Feed feedFragment = new NotificationFeed();
+
         if (eventId != null) {
             Bundle notificationBundle = new Bundle();
-            notificationBundle.putInt(getString(R.string.FEED_TYPE), FeedType.NOTIFICATION.ordinal());
             notificationBundle.putString(getString(R.string.event_id), eventId);
-            FeedFragment feedFragment = new FeedFragment();
             feedFragment.setArguments(notificationBundle);
             getSupportFragmentManager().beginTransaction().add(R.id.notification_feed, feedFragment).commit();
         } else {
@@ -43,7 +44,6 @@ public class MainActivity extends AppCompatActivity implements ClickableActivity
             // Notification feed fragment
             Bundle notificationBundle = new Bundle();
             notificationBundle.putInt(getString(R.string.FEED_TYPE), FeedType.NOTIFICATION.ordinal());
-            FeedFragment feedFragment = new FeedFragment();
             feedFragment.setArguments(notificationBundle);
             getSupportFragmentManager().beginTransaction().add(R.id.notification_feed, feedFragment).commit();
         }
